requirements.txt
asgiref==3.7.2
Django==5.0.3
djangorestframework==3.15.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1
